+++ app/Providers/AppServiceProvider.php
@@ -5,10 +5,13 @@ namespace App\Providers;
 use App\Models\Post;
 use App\Models\Setting;
 use App\Policies\PostPolicy;
+use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Support\Facades\View;
 use Illuminate\Support\ServiceProvider;
+use SocialiteProviders\LaravelPassport\LaravelPassportExtendSocialite;
+use SocialiteProviders\Manager\SocialiteWasCalled;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -25,6 +28,10 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot(): void
     {
+        Event::listen(function (SocialiteWasCalled $event): void {
+            $event->extendSocialite('laravelpassport', LaravelPassportExtendSocialite::class);
+        });
+
         Gate::policy(Post::class, PostPolicy::class);
 
         View::composer(['layouts.app', 'home', 'product', 'post'], function ($view) {