+++ bootstrap/providers.php
@@ -1,8 +1,8 @@
 <?php
- 
+
 use App\Providers\AppServiceProvider;
 use App\Providers\Filament\AdminPanelProvider;
- 
+
 return [
     AppServiceProvider::class,
     AdminPanelProvider::class,